+++ HISTORY.rst
@@ -23,8 +23,8 @@ Unreleased
 * ALCN-1039: Add suite_name remove invoice_address_id on webservice sale order web
 * ALCN-1039: Add esb_ref on deliver_carrier and update import data
 * Setup Cutoff Accrual
-
 * ALCN-1050: Add the time limit of order on the partner
+* ALCN-1057: Add IdRound in Customer XML export
 
 **Bugfixes**